# openapi.yaml
openapi: 3.0.3
info:
  title: Mountains Wonders - OpenAPI 3.0
  description: |- 
   Mountains wonders apis
  version: 1.0.0
...
paths:
  /auth/currentUser:
    get:
      tags:
        - auth
      summary: Retrieve the currentUser
      description: Retrieve the currentUser
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'          
            application/xml:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: user not logged
...
components:
  schemas:
    User:
      type: object
      properties:
        name:
          type: string
          example: giorgio
        email:
          type: string
          format: email
          example: example@gmail.com
        password:
          type: string
        isActive:
          type: boolean
        isAdmin:
          type: boolean
      xml:
        name: order
...